Refactor proposal-single tests.
