+++ contracts/cw-core/Cargo.toml
@@ -4,14 +4,6 @@ version = "0.2.0"
 authors = ["Zeke Medley <zekemedley@gmail.com>"]
 edition = "2018"
 
-exclude = [
-  # Those files are rust-optimizer artifacts. You might want to commit
-  # them for convenience but they should not be part of the source
-  # code publication.
-  "contract.wasm",
-  "hash.txt",
-]
-
 [lib]
 crate-type = ["cdylib", "rlib"]
 
@@ -20,13 +12,6 @@ crate-type = ["cdylib", "rlib"]
 backtraces = ["cosmwasm-std/backtraces"]
 # use library feature to disable all instantiate/execute/query exports
 library = []
-
-[package.metadata.scripts]
-optimize = """docker run --rm -v "$(pwd)":/code \
-  --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
-  --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-  cosmwasm/workspace-optimizer:0.12.6
-"""
 
 [dependencies]
 cosmwasm-std = { version = "1.0.0", features = ["ibc3"] }
