+++ contracts/cw-proposal-single/src/lib.rs
@@ -47,15 +47,10 @@ mod error;
 pub mod msg;
 pub mod proposal;
 pub mod query;
-#[cfg(test)]
-mod staking_tests;
 
 #[cfg(test)]
-mod mock_tests;
+mod testing;
 
 pub mod state;
-
-#[cfg(test)]
-mod tests;
 
 pub use crate::error::ContractError;
