+++ contracts/cw-proposal-single/src/testing/mod.rs
@@ -1,0 +1,9 @@
+mod contracts;
+mod do_votes;
+mod execute;
+mod instantiate;
+mod queries;
+
+mod tests;
+
+pub(crate) const CREATOR_ADDR: &str = "creator";
