+++ packages/cw-denom/Cargo.toml
@@ -4,7 +4,8 @@ version = "0.1.0"
 edition = "2021"
 authors = ["ekez ekez@withoutdoing.com"]
 repository = "https://github.com/DA0-DA0/dao-contracts"
-description = "Package for validation and handling of cw20 and native Cosmos SDK denominations."
+description = "A package for validation and handling of cw20 and native Cosmos SDK denominations."
+
 
 [dependencies]
 cosmwasm-std = "1.0"
